leave ca-certificates installed
A common use-case for the `mongo` image is to easily access the MongoDB CLI
where one is not natively installed. However, the `mongo` image currently
ships without `ca-certificates` which causes the CLI to fail when connecting
to MongoDB servers that require SSL.

The `ca-certificates` package is already installed during the build process
of this image, but it is removed during the build. This commit changes the
`Dockerfile` to leave `ca-certificates` installed. The size increase is
modest (~ 1 MB) but the increase in usefulness is vast.