Force temporary server to use unix socket only
This better matches the "initdb" behavior of the various SQL images, and avoids issues like running two instances in the same pod on different ports (and thus one of them failing to bind to 127.0.0.1:27017).

This is supported as of MongoDB 3.5+, and was discovered via https://stackoverflow.com/a/39901199/433558 -> https://jira.mongodb.org/browse/SERVER-9383 -> https://jira.mongodb.org/browse/SERVER-29403 (which was closed/fixed/implemented in 3.5.9).