Update `versions.sh` to filter targets correctly for pre-releases
This is relevant right now because 8.0.0-rc15 adds support for ubuntu 24.04 (noble), but 7.0 doesn't support it yet, so it's impossible to install the `noble` pre-release builds.

This fixes the calculation of which base image to use by cross-referencing this case.  While doing so, I centralized more of the logic for when/where we need that backwards version so we don't have to do the same weird calculations in multiple places / languages (with a new `dockerNeedsVersion` field in `versions.json`).

With this change, I was able to run `./versions.sh` and successfully build 8.0.0-rc15.